wrap beam_search function
