--- .idea/misc.xml
@@ -1,7 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project version="4">
-  <component name="PreferredVcsStorage">
-    <preferredVcsName>ApexVCS</preferredVcsName>
-  </component>
-  <component name="ProjectRootManager" version="2" project-jdk-name="Python 3.6.2 (E:\anaconda\python.exe)" project-jdk-type="Python SDK" />
-</project>
