+++ README.md
@@ -1,7 +1,4 @@
-=================================================更新===========================================================
-训练好的模型已经上传到百度云网盘，如果大家有需要可以前去下载。模型训练速度的话，CPU，16G 内存，一天即刻训练完成~~~
-
-链接：https://pan.baidu.com/s/1hrNxaSk 密码：d2sn
+# seq2seq_chatbot
 
 =================================================分割线，下面是正文===============================================
